toolkit/components/url-classifier/tests/mochitest/classifierFrame.html

<html>
<head>
<title></title>

<script type="text/javascript">

var scriptItem = "untouched";

function checkLoads() {
  // Make sure the javascript did not load.
  window.parent.is(scriptItem, "untouched", "Should not load bad javascript");

  // Make sure the css did not load.
  var elt = document.getElementById("styleCheck");
  var style = document.defaultView.getComputedStyle(elt, "");
  window.parent.isnot(style.visibility, "hidden", "Should not load bad css");

  elt = document.getElementById("styleBad");
  style = document.defaultView.getComputedStyle(elt, "");
  window.parent.isnot(style.visibility, "hidden", "Should not load bad css");

  elt = document.getElementById("styleImport");
  style = document.defaultView.getComputedStyle(elt, "");
  window.parent.isnot(style.visibility, "visible", "Should import clean css");

  // Call parent.loadTestFrame again to test classification metadata in HTTP
  // cache entries.
  if (window.parent.firstLoad) {
    window.parent.info("Reloading from cache...");
    window.parent.firstLoad = false;
    window.parent.loadTestFrame();
    return;
  }

  let dwu = window.parent.SpecialPowers.getDOMWindowUtils(window);
  let timer1 = window.setTimeout(function(){}, 0);
  window.parent.ok(!dwu.isTimeoutTracking(timer1),
                   "Timeout set from main script should not be considered as tracking");
  let timer2 = getTrackerTimeout();
  window.parent.ok(dwu.isTimeoutTracking(timer2),
                   "Timeout set from included script should be considered as tracking");
  window.clearTimeout(timer1);
  window.clearTimeout(timer2);

  // End (parent) test.
  window.parent.SimpleTest.finish();
}

</script>

<!-- Try loading from a malware javascript URI -->
<script type="text/javascript" src="http://malware.example.com/tests/toolkit/components/url-classifier/tests/mochitest/evil.js"></script>

<!-- Try loading from a tracker javascript URI -->
<script type="text/javascript" src="http://tracking.example.com/tests/toolkit/components/url-classifier/tests/mochitest/tracker.js"></script>

<!-- Try loading from an uwanted software css URI -->
<link rel="stylesheet" type="text/css" href="http://unwanted.example.com/tests/toolkit/components/url-classifier/tests/mochitest/evil.css"></link>

<!-- Try loading a marked-as-malware css through an @import from a clean URI -->
<link rel="stylesheet" type="text/css" href="import.css"></link>
</head>

<body onload="checkLoads()">
The following should not be hidden:
<div id="styleCheck">STYLE TEST</div>
<div id="styleBad">STYLE BAD</div>
<div id="styleImport">STYLE IMPORT</div>
</body>
</html>
